Move transfer project to team to componenet
