+++ src/lib/components/prompts/about/AboutProject.svelte
@@ -8,11 +8,12 @@
 	import Trash from '$lib/assets/Trash.svelte';
 	import TransferTeam from '$lib/components/dropdowns/projects/TransferTeam.svelte';
 	import InvitedTeamMembers from '$lib/components/projects/about/InvitedTeamMembers.svelte';
+	import TransferProjectToTeam from '$lib/components/projects/about/TransferProjectToTeam.svelte';
 	import NewTagsInput from '$lib/components/projects/edit/NewTagsInput.svelte';
 	import TagList from '$lib/components/projects/tags/TagList.svelte';
 	import Description from '$lib/components/text/Description.svelte';
 
-	import { currentProject } from '$lib/stores/project';
+	import { currentProject, userTeams } from '$lib/stores/project';
 	import { showManageInvitedPrompt } from '$lib/stores/ui';
 	import { supabase } from '$lib/supabase';
 	import lodash from 'lodash';
@@ -226,37 +227,9 @@
 
 		{#if !inEditMode}
 			<Description banner="" description={$currentProject.description} />
-			<div
-				class="relative mt-md flex items-center gap-sm overflow-visible font-medium text-grey-700 dark:text-grey-200 md:static"
-				bind:this={transferDropdownContainer}
-			>
-				{#if $currentProject.team}
-					<span>Part of</span>
-					<button
-						class="button--text m-0 flex items-center gap-sm p-0 md:relative"
-						on:click={() => (showTransferTeamDropdown = !showTransferTeamDropdown)}
-					>
-						{teamName}
-						<Down className="w-6 h-6 stroke-grey-700 dark:stroke-grey-200" />
-						{#if showTransferTeamDropdown}
-							<TransferTeam bind:visibility={showTransferTeamDropdown} />
-						{/if}
-					</button>
-				{:else}
-					<span>Transfer this project to a</span>
-					<button
-						class="button--text m-0 flex items-center gap-sm p-0 md:relative"
-						on:click={() => (showTransferTeamDropdown = !showTransferTeamDropdown)}
-					>
-						team
-						<Down className="w-6 h-6 stroke-grey-700 dark:stroke-grey-200" />
-						{#if showTransferTeamDropdown}
-							<TransferTeam bind:visibility={showTransferTeamDropdown} />
-						{/if}
-					</button>
-				{/if}
-			</div>
-
+			{#if $userTeams.length > 0}
+				<TransferProjectToTeam {teamName} />
+			{/if}
 			<InvitedTeamMembers invited_people={$currentProject.invited_people} />
 		{:else}
 			<label for="description-input" class="input--label mb-sm">Edit the project description</label>
